# core: Improve access tracking performance for read/write