--- a/core/src/main/java/org/dcache/nfs/v4/FileTracker.java
+++ b/core/src/main/java/org/dcache/nfs/v4/FileTracker.java
@@ -475,7 +475,7 @@ public class FileTracker {
         lock.lock();
         try {
 
-            switch (stateid.byteAt(11)) {
+            switch (stateid4.getType(stateid)) {
                 case Stateids.LOCK_STATE_ID:
                     NFS4State lockState = client.state(stateid);
                     stateid = lockState.getOpenState().stateid().getOpaque();