--- a/core/src/main/java/org/dcache/nfs/v4/NFS4State.java
+++ b/core/src/main/java/org/dcache/nfs/v4/NFS4State.java
@@ -68,7 +68,7 @@ public class NFS4State {
     }
 
     public void bumpSeqid() {
-        ++_stateid.seqid;
+        _stateid.bumpSeqid();
     }
 
     public stateid4 stateid() {