--- a/core/src/main/java/org/dcache/nfs/v4/OperationCOMMIT.java
+++ b/core/src/main/java/org/dcache/nfs/v4/OperationCOMMIT.java
@@ -28,6 +28,7 @@ import org.dcache.nfs.v4.xdr.COMMIT4resok;
 import org.dcache.nfs.v4.xdr.nfs_argop4;
 import org.dcache.nfs.v4.xdr.nfs_opnum4;
 import org.dcache.nfs.v4.xdr.nfs_resop4;
+import org.dcache.nfs.v4.xdr.stateid4;
 import org.dcache.nfs.vfs.Inode;
 
 public class OperationCOMMIT extends AbstractNFSv4Operation {
@@ -42,8 +43,11 @@ public class OperationCOMMIT extends AbstractNFSv4Operation {
         final COMMIT4res res = result.opcommit;
         Inode inode = context.currentInode();
 
+        stateid4 stateid = Stateids.getCurrentStateidIfNeeded(context, _args.opwrite.stateid);
+
         _args.opcommit.offset.checkOverflow(_args.opcommit.count.value, "offset + length overflow");
-        context.getFs().commit(inode, _args.opcommit.offset.value, _args.opcommit.count.value);
+        context.getFs().commit(stateid.getOpaque(), inode, _args.opcommit.offset.value,
+                _args.opcommit.count.value);
 
         res.resok4 = new COMMIT4resok();
         res.resok4.writeverf = context.getRebootVerifier();