@kofemann 2025-07-14
There are too many changes. Some make sense, some don't. Cleanups, like `new stateid` => `stateid.clone()`, are good. `byteAt` should be `stateid.type()`. The vfs layer is extra designed to be thin. If OPEN/READ/WRITE/CLOSE must pass extra information to the custom filesystem, then those operations can be overridden.