@kohlschuetter 2025-07-14
I'm not sure I follow.

What can be removed? We're not going to expose `byte[]` or `stateid4` directly to the VFS API, so deep cuts have to be made.

How can operations be overridden by the VFS if that information is not exposed otherwise?

Which changes don't make sense?