@kohlschuetter 2025-07-14
@kofemann I've cut a few changes, but that's as much as I think we can cut.